added user id while karta deletion

+++ src/app/components/dashboard/dashboard.component.ts
@@ -160,7 +160,7 @@ export class DashboardComponent implements OnInit {
   deleteKarta(id: string) {
     const result = confirm('Are you sure you want to delete this karta?');
     if (result) {
-      this._kartaService.deleteKarta({ kartaId: id }).subscribe(
+      this._kartaService.deleteKarta({ kartaId: id, userId: this._commonService.getUserId() }).subscribe(
         (response: any) => {
           this._commonService.successToaster("Karta deleted successfully!");
           this.getAllKartas();
