+++ src/app/components/karta/all-kartas/all-kartas.component.ts
@@ -146,7 +146,7 @@ export class AllKartasComponent implements OnInit {
   deleteKarta(id: string) {
     const result = confirm("Are you sure you want to delete this karta?");
     if (result) {
-      this._kartaService.deleteKarta({ kartaId: id }).subscribe(
+      this._kartaService.deleteKarta({ kartaId: id, userId: this._commonService.getUserId() }).subscribe(
         (response: any) => {
           this._commonService.successToaster("Karta deleted successfully!");
           this.getAllKartas();
